Feat: Add dynamic title to customer support screen
The customer support screen now receives its title from the Get arguments. If no title is provided, it defaults to 'Customer Support'. The title is displayed in the app bar of the screen. The `title` parameter was removed from `CustomerSupportPage` and `CustomerSupportScreen` since the `title` now comes from the arguments.